Refactor: migrate API response structures to new files and implement Metadata handling
