+++ pkg/api/success.go
@@ -1,0 +1,21 @@
+package api
+
+import "time"
+
+type SuccessResponse struct {
+	StatusCode  int      `json:"statusCode,omitempty" bson:"statusCode,omitempty"`
+	RequestId   string   `json:"requestId,omitempty" bson:"requestId,omitempty"`
+	MetaData    Metadata `json:"metaData,omitempty" bson:"metaData,omitempty"`
+	SuccessCode string   `json:"successCode,omitempty" bson:"successCode,omitempty"`
+	Message     string   `json:"message,omitempty" bson:"message,omitempty"`
+}
+
+func CreateSuccess(statusCode int, successCode string, message string, metadata Metadata) SuccessResponse {
+	metadata.Timestamp = time.Now().Unix()
+	return SuccessResponse{
+		StatusCode:  statusCode,
+		MetaData:    metadata,
+		SuccessCode: successCode,
+		Message:     message,
+	}
+}
